+++ programs/client/Suggest.h
@@ -4,7 +4,7 @@
 
 #include <Client/Connection.h>
 #include <IO/ConnectionTimeouts.h>
-#include <common/LineReader.h>
+#include <Common/LineReader.h>
 #include <thread>
