+++ src/Common/LineReader.h
